+++ http-cache/src/managers/cacache.rs
@@ -4,7 +4,6 @@ use crate::{CacheManager, HttpResponse, Result};
 
 use http_cache_semantics::CachePolicy;
 use serde::{Deserialize, Serialize};
-use url::Url;
 
 /// Implements [`CacheManager`] with [`cacache`](https://github.com/zkat/cacache-rs) as the backend.
 #[cfg_attr(docsrs, doc(cfg(feature = "manager-cacache")))]
@@ -26,10 +25,6 @@ struct Store {
     policy: CachePolicy,
 }
 
-fn req_key(method: &str, url: &Url) -> String {
-    format!("{method}:{url}")
-}
-
 #[allow(dead_code)]
 impl CACacheManager {
     /// Clears out the entire cache.
@@ -43,33 +38,30 @@ impl CACacheManager {
 impl CacheManager for CACacheManager {
     async fn get(
         &self,
-        method: &str,
-        url: &Url,
+        cache_key: &str,
     ) -> Result<Option<(HttpResponse, CachePolicy)>> {
-        let store: Store =
-            match cacache::read(&self.path, &req_key(method, url)).await {
-                Ok(d) => bincode::deserialize(&d)?,
-                Err(_e) => {
-                    return Ok(None);
-                }
-            };
+        let store: Store = match cacache::read(&self.path, cache_key).await {
+            Ok(d) => bincode::deserialize(&d)?,
+            Err(_e) => {
+                return Ok(None);
+            }
+        };
         Ok(Some((store.response, store.policy)))
     }
 
     async fn put(
         &self,
-        method: &str,
-        url: &Url,
+        cache_key: String,
         response: HttpResponse,
         policy: CachePolicy,
     ) -> Result<HttpResponse> {
         let data = Store { response: response.clone(), policy };
         let bytes = bincode::serialize(&data)?;
-        cacache::write(&self.path, &req_key(method, url), bytes).await?;
+        cacache::write(&self.path, cache_key, bytes).await?;
         Ok(response)
     }
 
-    async fn delete(&self, method: &str, url: &Url) -> Result<()> {
-        Ok(cacache::remove(&self.path, &req_key(method, url)).await?)
+    async fn delete(&self, cache_key: &str) -> Result<()> {
+        Ok(cacache::remove(&self.path, cache_key).await?)
     }
 }